clean unused proto and fix uint64 overflow bug

+++ privval/file.go
@@ -315,7 +315,7 @@ func (pv *FilePV) signOracleVote(vote *oracleproto.GossipVote) error {
 	if err != nil {
 		return err
 	}
-	vote.SignedTimestamp = uint64(time.Now().Unix())
+	vote.SignedTimestamp = time.Now().Unix()
 	vote.Signature = sig
 
 	return nil
